src/services/instagram_intelligence/api_server.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
Instagram Intelligence + AI Vision Analysis API
Mobile proxy-based Instagram scraping with AI vision analysis
"""

import os
import json
import time
import hashlib
import requests
from datetime import datetime
from typing import List, Dict, Optional
from dataclasses import dataclass
from flask import Flask, request, jsonify
from PIL import Image
import io

app = Flask(__name__)

# Configuration
PROXY_SX_API_KEY = os.environ.get('PROXY_SX_API_KEY', '')
OPENAI_API_KEY = os.environ.get('OPENAI_API_KEY', '')
PROXY_BASE_URL = 'https://proxies.sx/api/v1'

@dataclass
class InstagramProfile:
    """Represents an Instagram profile"""
    username: str
    full_name: str
    biography: str
    followers: int
    following: int
    posts_count: int
    is_verified: bool
    category: str
    profile_pic_url: str
    engagement_rate: float

@dataclass
class InstagramPost:
    """Represents an Instagram post"""
    id: str
    caption: str
    likes: int
    comments: int
    timestamp: str
    media_url: str
    media_type: str

class InstagramScraper:
    """Instagram scraping with mobile proxies"""
    
    def __init__(self, api_key: str = ''):
        self.api_key = api_key
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 16_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.0 Mobile/15E148 Safari/604.1',
            'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8',
            'Accept-Language': 'en-US,en;q=0.5',
        })
    
    def get_proxy(self) -> Optional[Dict]:
        """Get a mobile proxy"""
        if not self.api_key:
            return None
        
        try:
            resp = requests.get(
                f'{PROXY_BASE_URL}/allocate',
                params={'country': 'US', 'type': 'mobile'},
                headers={'Authorization': f'Bearer {self.api_key}'},
                timeout=10
            )
            if resp.status_code == 200:
                return resp.json()
        except:
            pass
        
        return None
    
    def get_profile(self, username: str) -> Optional[InstagramProfile]:
        """Get Instagram profile data"""
        url = f'https://www.instagram.com/{username}/?__a=1'
        
        proxy = self.get_proxy()
        
        try:
            if proxy:
                resp = requests.get(url, proxies=proxy, timeout=30)
            else:
                resp = self.session.get(url, timeout=30)
            
            if resp.status_code == 200:
                data = resp.json()
                return self._parse_profile(data, username)
            
        except Exception as e:
            print(f"Profile fetch error: {e}")
        
        # Return demo data
        return InstagramProfile(
            username=username,
            full_name='Demo User',
            biography='Demo Instagram account',
            followers=10000,
            following=500,
            posts_count=100,
            is_verified=False,
            category='Personal',
            profile_pic_url='',
            engagement_rate=5.2
        )
    
    def get_posts(self, username: str, limit: int = 12) -> List[InstagramPost]:
        """Get recent posts from a profile"""
        url = f'https://www.instagram.com/{username}/?__a=1'
        
        proxy = self.get_proxy()
        
        try:
            if proxy:
                resp = requests.get(url, proxies=proxy, timeout=30)
            else:
                resp = self.session.get(url, timeout=30)
            
            if resp.status_code == 200:
                data = resp.json()
                return self._parse_posts(data, limit)
        except Exception as e:
            print(f"Posts fetch error: {e}")
        
        return []
    
    def _parse_profile(self, data: dict, username: str) -> Optional[InstagramProfile]:
        """Parse profile data from Instagram response"""
        try:
            graphql = data.get('graphql', {})
            user = graphql.get('user', {})
            
            return InstagramProfile(
                username=username,
                full_name=user.get('full_name', ''),
                biography=user.get('biography', ''),
                followers=user.get('edge_followed_by', {}).get('count', 0),
                following=user.get('edge_follow', {}).get('count', 0),
                posts_count=user.get('edge_owner_media', {}).get('count', 0),
                is_verified=user.get('is_verified', False),
                category=user.get('category_enum_name', ''),
                profile_pic_url=user.get('profile_pic_url_hd', ''),
                engagement_rate=self._calc_engagement(user)
            )
        except:
            return None
    
    def _calc_engagement(self, user: dict) -> float:
        """Calculate engagement rate"""
        posts = user.get('edge_owner_media', {}).get('edges', [])
        if not posts:
            return 0.0
        
        total_likes = 0
        total_comments = 0
        followers = user.get('edge_followed_by', {}).get('count', 1)
        
        for post in posts[:10]:
            node = post.get('node', {})
            total_likes += node.get('edge_media_preview_like', {}).get('count', 0)
            total_comments += node.get('edge_media_to_comment', {}).get('count', 0)
        
        avg_engagement = (total_likes + total_comments) / 10
        return round((avg_engagement / followers) * 100, 2)
    
    def _parse_posts(self, data: dict, limit: int) -> List[InstagramPost]:
        """Parse posts from Instagram response"""
        posts = []
        
        try:
            graphql = data.get('graphql', {})
            user = graphql.get('user', {})
            edges = user.get('edge_owner_media', {}).get('edges', [])[:limit]
            
            for edge in edges:
                node = edge.get('node', {})
                posts.append(InstagramPost(
                    id=node.get('id', ''),
                    caption=node.get('edge_media_to_caption', {}).get('edges', [{}])[0].get('node', {}).get('text', ''),
                    likes=node.get('edge_media_preview_like', {}).get('count', 0),
                    comments=node.get('edge_media_to_comment', {}).get('count', 0),
                    timestamp=node.get('taken_at_timestamp', ''),
                    media_url=node.get('display_url', ''),
                    media_type=node.get('__typename', 'Image')
                ))
        except Exception as e:
            print(f"Parse posts error: {e}")
        
        return posts


class AIVisionAnalyzer:
    """AI vision analysis for Instagram content"""
    
    def __init__(self, api_key: str = ''):
        self.api_key = api_key
    
    def analyze_images(self, image_urls: List[str]) -> Dict:
        """Analyze images using AI vision model"""
        
        # Demo analysis if no API key
        if not self.api_key:
            return self._demo_analysis()
        
        # Real implementation would use OpenAI/Anthropic
        try:
            # Placeholder for AI vision integration
            pass
        except:
            pass
        
        return self._demo_analysis()
    
    def _demo_analysis(self) -> Dict:
        """Return demo analysis"""
        return {
            'content_themes': ['lifestyle', 'travel'],
            'content_style': 'professional_photography',
            'brand_safety_score': 92,
            'content_consistency': 'high',
            'account_type': 'influencer',
            'niche': 'travel_lifestyle',
            'confidence': 0.94,
            'signals': [
                'professional_quality_images',
                'consistent_aesthetic',
                'engagement_pattern_matches_organic'
            ],
            'sentiment': {
                'overall': 'positive',
                'breakdown': {'positive': 72, 'neutral': 20, 'negative': 8},
                'themes': ['aspirational', 'happy', 'adventurous']
            },
            'engagement_prediction': {
                'predicted_likes': 15000,
                'predicted_comments': 500,
                'confidence': 0.85
            }
        }
    
    def detect_account_type(self, profile: InstagramProfile, posts: List[InstagramPost]) -> Dict:
        """Detect account type using AI analysis"""
        
        return {
            'account_type': 'influencer',
            'niche': 'travel_lifestyle',
            'confidence': 0.94,
            'signals': [
                'professional_quality_images',
                'consistent_aesthetic',
                'high_engagement_rate',
                'regular_posting_schedule'
            ]
        }


# Initialize components
scraper = InstagramScraper(PROXY_SX_API_KEY)
ai_analyzer = AIVisionAnalyzer(OPENAI_API_KEY)


# ============== API Endpoints ==============

@app.route('/health', methods=['GET'])
def health():
    """Health check"""
    return jsonify({
        'status': 'ok',
        'timestamp': datetime.now().isoformat()
    })

@app.route('/api/instagram/profile/<username>', methods=['GET'])
def get_profile(username: str):
    """Get Instagram profile data"""
    profile = scraper.get_profile(username)
    
    if not profile:
        return jsonify({'error': 'Profile not found'}), 404
    
    return jsonify({
        'username': profile.username,
        'full_name': profile.full_name,
        'biography': profile.biography,
        'followers': profile.followers,
        'following': profile.following,
        'posts_count': profile.posts_count,
        'is_verified': profile.is_verified,
        'category': profile.category,
        'engagement_rate': profile.engagement_rate,
        'profile_pic_url': profile.profile_pic_url
    })

@app.route('/api/instagram/posts/<username>', methods=['GET'])
def get_posts(username: str):
    """Get recent posts from a profile"""
    limit = min(int(request.args.get('limit', 12)), 50)
    posts = scraper.get_posts(username, limit)
    
    return jsonify({
        'username': username,
        'posts': [p.__dict__ for p in posts],
        'count': len(posts)
    })

@app.route('/api/instagram/analyze/<username>', methods=['GET'])
def analyze_account(username: str):
    """Full AI analysis of an account"""
    profile = scraper.get_profile(username)
    posts = scraper.get_posts(username, 12)
    
    if not profile:
        return jsonify({'error': 'Profile not found'}), 404
    
    # Get AI analysis
    ai_analysis = ai_analyzer.detect_account_type(profile, posts)
    
    return jsonify({
        'username': username,
        'profile': {
            'followers': profile.followers,
            'following': profile.following,
            'posts_count': profile.posts_count,
            'engagement_rate': profile.engagement_rate,
            'is_verified': profile.is_verified,
            'category': profile.category
        },
        'ai_analysis': ai_analysis,
        'analyzed_at': datetime.now().isoformat()
    })


# ============== Main ==============

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 5000))
    debug = os.environ.get('DEBUG', 'false').lower() == 'true'
    
    print('=' * 60)
    print('Instagram Intelligence + AI Vision API')
    print('=' * 60)
    print(f'Starting on port {port}...')
    print(f'Proxy API Key: {"Set" if PROXY_SX_API_KEY else "Not set (demo mode)"}')
    print(f'OpenAI API Key: {"Set" if OPENAI_API_KEY else "Not set (demo mode)"}')
    print()
    
    app.run(host='0.0.0.0', port=port, debug=debug)
